+++ mindmeld/system_entity_recognizer.py
@@ -22,7 +22,7 @@ import requests
 
 from .components._config import (DEFAULT_DUCKLING_URL, get_system_entity_url_config,
                                  is_duckling_configured)
-from .components.request import validate_language_code, validate_locale_code, validate_timestamp
+from .components.schemas import validate_language_code, validate_locale_code, validate_timestamp
 from .core import Entity, QueryEntity, Span, _sort_by_lowest_time_grain
 from .exceptions import MindMeldError, SystemEntityResolutionError
