+++ tests/test_app_manager_async.py
@@ -10,7 +10,6 @@ Tests for app_manager module.
 These tests apply only when async/await are supported.
 """
 import pytest
-from mindmeld.components.request import dialogue_response_schema
 
 
 @pytest.mark.asyncio
@@ -19,6 +18,6 @@ async def test_parse(async_kwik_e_mart_app):
     fields = {"params", "request", "dialogue_state", "directives", "history"}
     nested_fields = {"domain", "intent", "entities"}
     for field in fields:
-        assert field in dialogue_response_schema.dump(response).keys()
+        assert field in dict(response).keys()
     for field in nested_fields:
-        assert field in dialogue_response_schema.dump(response)['request'].keys()
+        assert field in dict(response)['request'].keys()
